+++ chpl/chpl-api/e2e/script/newman-run.js
@@ -378,6 +378,16 @@ const surveillanceControllerTests = {
   },
 };
 
+const conformanceMethodsControllerTests = {
+  ...commonOptions,
+  collection: collection_path + '/conformance-methods-controller.postman_collection.json',
+  reporter: {
+    junit: {
+      export: reports_path + '/conformance-methods-controller-tests.xml',
+    },
+  },
+};     
+      
 const searchQuestionableActivityControllerTests = {
   ...commonOptions,
   collection: collection_path + '/search-questionable-activity-controller.postman_collection.json',
@@ -394,6 +404,16 @@ const announcementsControllerTests = {
   reporter: {
     junit: {
       export: reports_path + '/announcements-controller-tests.xml',
+    },
+  },
+};
+
+const reportDataCriteriaUpToDateControllerTests = {
+  ...commonOptions,
+  collection: collection_path + '/report-data-criteria-up-to-date.postman_collection.json',
+  reporter: {
+    junit: {
+      export: reports_path + '/report-data-criteria-up-to-date-controller-tests.xml',
     },
   },
 };
@@ -435,8 +455,10 @@ const jobs = [
   cb => newman.run(meauresControllerTests,cb),
   cb => newman.run(targetedUsersControllerTests,cb),
   cb => newman.run(surveillanceControllerTests,cb),
+  cb => newman.run(conformanceMethodsControllerTests,cb),
   cb => newman.run(searchQuestionableActivityControllerTests,cb),
   cb => newman.run(announcementsControllerTests,cb),
+  cb => newman.run(reportDataCriteriaUpToDateControllerTests,cb),
   ];
 
 const responseCallback = (err) => {