+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/scheduler/job/updatedcriteriastatusreport/UpdatedCriteriaStatusReportSheet.java
@@ -1,9 +1,7 @@
 package gov.healthit.chpl.scheduler.job.updatedcriteriastatusreport;
 import java.time.LocalDate;
 import java.util.List;
-import java.util.stream.Collectors;
 
-import org.apache.commons.collections.CollectionUtils;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.ss.util.CellUtil;
@@ -13,14 +11,10 @@ import org.apache.poi.xssf.usermodel.XSSFSheet;
 import org.springframework.stereotype.Component;
 
 import gov.healthit.chpl.certificationCriteria.CertificationCriterion;
-import gov.healthit.chpl.dao.CertificationStatusDAO;
-import gov.healthit.chpl.dao.statistics.SummaryStatisticsDAO;
-import gov.healthit.chpl.domain.CertificationStatus;
-import gov.healthit.chpl.scheduler.job.summarystatistics.data.StatisticsSnapshot;
-import gov.healthit.chpl.util.CertificationStatusUtil;
+import gov.healthit.chpl.report.criteriauptodate.CriteriaUpToDateReport;
+import gov.healthit.chpl.report.criteriauptodate.CriteriaUpToDateReportService;
 import gov.healthit.chpl.util.Util;
 import lombok.extern.log4j.Log4j2;
-import one.util.streamex.StreamEx;
 
 @Log4j2(topic = "updatedCriteriaStatusReportEmailJobLogger")
 @Component
@@ -31,16 +25,10 @@ public class UpdatedCriteriaStatusReportSheet {
 
     private static final Integer DESCRIPTIONS_COL_IDX = 0;
 
-    private UpdatedCriterionStatusReportDao updatedCriteriaStatusReportDao;
-    private SummaryStatisticsDAO summaryStatisticsDao;
-    private List<CertificationStatus> allCertificationStatuses;
+    private CriteriaUpToDateReportService reportService;
 
-    public UpdatedCriteriaStatusReportSheet(UpdatedCriterionStatusReportDao updatedCriteriaStatusReportDao,
-            SummaryStatisticsDAO summaryStatisticsDao,
-            CertificationStatusDAO certificationStatusDao) {
-        this.updatedCriteriaStatusReportDao = updatedCriteriaStatusReportDao;
-        this.summaryStatisticsDao = summaryStatisticsDao;
-        this.allCertificationStatuses = certificationStatusDao.findAll();
+    public UpdatedCriteriaStatusReportSheet(CriteriaUpToDateReportService reportService) {
+        this.reportService = reportService;
     }
 
     public void generateSheetForCriteriaOnDates(CertificationCriterion criterion, List<LocalDate> reportDates, Workbook workbook) {
@@ -52,11 +40,10 @@ public class UpdatedCriteriaStatusReportSheet {
 
         for (int i = UpdatedCriteriaStatusReportWorkbook.TOTAL_NUMBER_OF_MONTHS; i >= 1; --i) {
             LocalDate actualReportDay = reportDates.get(i - 1);
-            List<UpdatedCriterionStatusReport> criterionStatusReports = getCriterionStatusReportsForDateAndCriterion(actualReportDay, criterion);
-            StatisticsSnapshot statisticsSnapshot = getSummaryStatisticsSnapshotForDate(actualReportDay);
+            List<CriteriaUpToDateReport> reports = reportService.getAllCriteriaUpToDateReports(actualReportDay);
 
-            long totalActiveListingsWithCriterion = calculateActiveListingsWithCriterionCount(statisticsSnapshot, criterion);
-            long totalListingsRequiringUpdates = calculateListingsRequiringUpdatesCount(criterionStatusReports);
+            long totalActiveListingsWithCriterion = calculateActiveListingsWithCriterionCount(reports, criterion);
+            long totalListingsRequiringUpdates = calculateListingsRequiringUpdatesCount(reports, criterion);
 
             CellUtil.getCell(CellUtil.getRow(DATE_ROW_IDX, sheet), i).setCellValue(actualReportDay);
             CellUtil.getCell(CellUtil.getRow(REQUIRES_UPDATE_ROW_IDX, sheet), i).setCellValue(totalListingsRequiringUpdates);
@@ -64,34 +51,28 @@ public class UpdatedCriteriaStatusReportSheet {
         }
     }
 
-    private long calculateActiveListingsWithCriterionCount(StatisticsSnapshot statisticsSnapshot, CertificationCriterion criterion) {
-        if (statisticsSnapshot == null
-                || CollectionUtils.isEmpty(statisticsSnapshot.getAttestedCriterionStatistics())) {
-            LOGGER.warn("No attested criterion statistics were found in the statistics snapshot");
+    private long calculateActiveListingsWithCriterionCount(List<CriteriaUpToDateReport> reports, CertificationCriterion criterion) {
+        CriteriaUpToDateReport reportForCriterion = reports.stream()
+                .filter(report -> report.getCriterion().getId().equals(criterion.getId()))
+                .findAny().orElse(null);
+        if (reportForCriterion == null) {
+            LOGGER.error("Unable to calculate active listings with criterion " + Util.formatCriteriaNumber(criterion));
             return 0;
         }
-        List<Long> activeStatusIds = allCertificationStatuses.stream()
-                .filter(certStatus -> CertificationStatusUtil.getActiveStatusNames().contains(certStatus.getName()))
-                .map(certStatus -> certStatus.getId())
-                .collect(Collectors.toList());
-
-        long countOfListingsInActiveStatusWithCriteria = statisticsSnapshot.getAttestedCriterionStatistics().stream()
-            .filter(stat -> stat.getCertificationCriterionId().equals(criterion.getId())
-                    && activeStatusIds.contains(stat.getListingStatusId()))
-            .map(stat -> stat.getListingIds().size())
-            .collect(Collectors.summingInt(Integer::intValue));
-        return countOfListingsInActiveStatusWithCriteria;
+        return reportForCriterion.getActiveListingsAttestingToCriterionCount();
     }
 
-    private long calculateListingsRequiringUpdatesCount(List<UpdatedCriterionStatusReport> reports) {
-        if (CollectionUtils.isEmpty(reports)) {
-            LOGGER.warn("No updated criteria status reports were found");
+    private long calculateListingsRequiringUpdatesCount(List<CriteriaUpToDateReport> reports, CertificationCriterion criterion) {
+        CriteriaUpToDateReport reportForCriterion = reports.stream()
+                .filter(report -> report.getCriterion().getId().equals(criterion.getId()))
+                .findAny().orElse(null);
+
+        if (reportForCriterion == null) {
+            LOGGER.error("Unable to calculate listings requiring update count for criterion " + Util.formatCriteriaNumber(criterion));
             return 0;
         }
 
-        return StreamEx.of(reports)
-                .distinct(UpdatedCriterionStatusReport::getCertifiedProductId)
-                .count();
+        return reportForCriterion.getActiveListingsAttestingToCriterionCount() - reportForCriterion.getActiveListingsUpToDateOnCriterionCount();
     }
 
     private void updateChartTitle(Sheet sheet, CertificationCriterion criterion) {
@@ -102,16 +83,6 @@ public class UpdatedCriteriaStatusReportSheet {
                 charts.get(0).setTitleText(Util.formatCriteriaNumber(criterion) + " Up-to-Date Progress");
             }
         }
-    }
-
-    private List<UpdatedCriterionStatusReport> getCriterionStatusReportsForDateAndCriterion(LocalDate reportDate, CertificationCriterion criterion) {
-        return updatedCriteriaStatusReportDao.getUpdatedCriterionStatusReportsByDay(reportDate).stream()
-                .filter(report -> report.getCertificationCriterion().getId().equals(criterion.getId()))
-                .collect(Collectors.toList());
-    }
-
-    private StatisticsSnapshot getSummaryStatisticsSnapshotForDate(LocalDate reportDate) {
-        return summaryStatisticsDao.getSummaryStatistics(reportDate);
     }
 
     private Sheet addWorksheetForCriteria(CertificationCriterion criterion, Workbook workbook) {